move actions up a level
