+++ src/client/action.rs
@@ -1,42 +1,2 @@
-//! Action builder types.
-
-pub(crate) mod list_databases;
-pub(crate) mod perf;
-pub(crate) mod session;
-pub(crate) mod shutdown;
-pub(crate) mod watch;
-
-pub use list_databases::ListDatabases;
-pub use perf::WarmConnectionPool;
-pub use session::StartSession;
-pub use shutdown::Shutdown;
-pub use watch::Watch;
-
-macro_rules! option_setters {
-    (
-        $opt_field:ident: $opt_field_ty:ty;
-        $(
-            $(#[$($attrss:tt)*])*
-            $opt_name:ident: $opt_ty:ty,
-        )+
-    ) => {
-        fn options(&mut self) -> &mut $opt_field_ty {
-            self.$opt_field.get_or_insert_with(<$opt_field_ty>::default)
-        }
-
-        #[cfg(test)]
-        pub(crate) fn with_options(mut self, value: impl Into<Option<$opt_field_ty>>) -> Self {
-            self.options = value.into();
-            self
-        }
-
-        $(
-            $(#[$($attrss)*])*
-            pub fn $opt_name(mut self, value: $opt_ty) -> Self {
-                self.options().$opt_name = Some(value);
-                self
-            }
-        )+
-    };
-}
-use option_setters;
+mod perf;
+mod shutdown;
